DamagoFirebase/functions/requirements.txt
firebase_functions~=0.1.0
nanoid
firebase-admin
google-cloud-firestore
